# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwarmCast API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateAnalysisBody {
  title: string;
  text: string;
}

export interface CreateAnalysisResponse {
  id: string;
  status: string;
}

export interface AnalysisSummary {
  id: string;
  title: string;
  status: string;
  createdAt: string;
  /** @nullable */
  avgSentiment?: number | null;
  /** @nullable */
  dominantEmotion?: string | null;
  /** @nullable */
  riskLevel?: string | null;
}

export interface FactCheckItem {
  claim: string;
  status: string;
  note: string;
}

export interface FactCheckResult {
  objectiveAssessment: string;
  accuracyRating: string;
  factCheckItems: FactCheckItem[];
}

export interface ExaResult {
  title: string;
  url: string;
  snippet: string;
  /** @nullable */
  publishedDate?: string | null;
}

export interface AlignmentWord {
  word: string;
  start: number;
  end: number;
}

export interface AlignmentData {
  words: AlignmentWord[];
}

export interface Persona {
  id: number;
  analysisId: string;
  /** @nullable */
  personaName?: string | null;
  /** @nullable */
  personaType?: string | null;
  /** @nullable */
  background?: string | null;
  /** @nullable */
  mbti?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  accent?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  voiceStyle?: string | null;
  /** @nullable */
  activityLevel?: number | null;
  /** @nullable */
  influenceWeight?: number | null;
  /** @nullable */
  sentimentBias?: number | null;
  /** @nullable */
  initialReaction?: string | null;
  /** @nullable */
  finalOpinion?: string | null;
  /** @nullable */
  initialSentiment?: number | null;
  /** @nullable */
  finalSentiment?: number | null;
  /** @nullable */
  beliefStance?: number | null;
  /** @nullable */
  beliefConfidence?: number | null;
  /** @nullable */
  keyConcern?: string | null;
  /** @nullable */
  wouldShare?: boolean | null;
  /** @nullable */
  platformPreference?: string | null;
  /** @nullable */
  voiceId?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  hasAudio?: boolean | null;
  /** @nullable */
  voiceScore?: number | null;
  /** @nullable */
  rank?: number | null;
  alignmentData?: AlignmentData | null;
}

export interface ForecastPoint {
  id: number;
  analysisId: string;
  /** @nullable */
  hourOffset?: number | null;
  /** @nullable */
  sentiment?: number | null;
  /** @nullable */
  confidenceLow?: number | null;
  /** @nullable */
  confidenceHigh?: number | null;
  /** @nullable */
  isForecast?: boolean | null;
}

export interface ProblemSegment {
  quote: string;
  triggeredBy: string[];
  reason: string;
}

export interface MontageTimelineEntry {
  /** @nullable */
  personaId?: number | null;
  personaName: string;
  startSec: number;
  endSec: number;
  script: string;
  /** @nullable */
  words?: AlignmentWord[] | null;
}

export interface SwarmSignedUrlResponse {
  signedUrl: string;
}

export interface FullAnalysis {
  id: string;
  title: string;
  inputText: string;
  status: string;
  createdAt: string;
  /** @nullable */
  avgSentiment?: number | null;
  /** @nullable */
  dominantEmotion?: string | null;
  /** @nullable */
  riskLevel?: string | null;
  /** @nullable */
  viralPotential?: number | null;
  /** @nullable */
  consensusForming?: boolean | null;
  /** @nullable */
  swarmSummary?: string | null;
  /** @nullable */
  marketQuestion?: string | null;
  /** @nullable */
  marketProbability?: number | null;
  factCheck?: FactCheckResult | null;
  /** @nullable */
  exaResults?: ExaResult[] | null;
  /** @nullable */
  keyThemes?: string[] | null;
  /** @nullable */
  narrativeFractures?: string[] | null;
  /** @nullable */
  montageUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  contentSuggestions?: string[] | null;
  /** @nullable */
  problemSegments?: ProblemSegment[] | null;
  /** @nullable */
  montageTimeline?: MontageTimelineEntry[] | null;
  personas: Persona[];
  forecastPoints: ForecastPoint[];
}