# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwarmCast API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analyses
    description: SwarmCast analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List recent analyses
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisSummary"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create a new analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAnalysisResponse"
        "400":
          description: Invalid input

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get full analysis result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullAnalysis"
        "404":
          description: Not found

  /analyses/{id}/stream:
    get:
      operationId: streamAnalysis
      tags: [analyses]
      summary: SSE stream for analysis pipeline events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string

  /analyses/{id}/swarm-signed-url:
    get:
      operationId: getSwarmSignedUrl
      tags: [analyses]
      summary: Get signed URL for SwarmCast ConvAI session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signed URL for ElevenLabs ConvAI session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwarmSignedUrlResponse"
        "404":
          description: Analysis not found or agent not ready

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CreateAnalysisBody:
      type: object
      properties:
        title:
          type: string
        text:
          type: string
      required:
        - title
        - text

    CreateAnalysisResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
      required:
        - id
        - status

    AnalysisSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        createdAt:
          type: string
        avgSentiment:
          type: ["number", "null"]
        dominantEmotion:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
      required:
        - id
        - title
        - status
        - createdAt

    FactCheckItem:
      type: object
      properties:
        claim:
          type: string
        status:
          type: string
        note:
          type: string
      required:
        - claim
        - status
        - note

    FactCheckResult:
      type: object
      properties:
        objectiveAssessment:
          type: string
        accuracyRating:
          type: string
        factCheckItems:
          type: array
          items:
            $ref: "#/components/schemas/FactCheckItem"
      required:
        - objectiveAssessment
        - accuracyRating
        - factCheckItems

    ExaResult:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
        publishedDate:
          type: ["string", "null"]
      required:
        - title
        - url
        - snippet

    AlignmentWord:
      type: object
      properties:
        word:
          type: string
        start:
          type: number
        end:
          type: number
      required:
        - word
        - start
        - end

    AlignmentData:
      type: object
      properties:
        words:
          type: array
          items:
            $ref: "#/components/schemas/AlignmentWord"
      required:
        - words

    Persona:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: string
        personaName:
          type: ["string", "null"]
        personaType:
          type: ["string", "null"]
        background:
          type: ["string", "null"]
        mbti:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        country:
          type: ["string", "null"]
        accent:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        voiceStyle:
          type: ["string", "null"]
        activityLevel:
          type: ["number", "null"]
        influenceWeight:
          type: ["number", "null"]
        sentimentBias:
          type: ["number", "null"]
        initialReaction:
          type: ["string", "null"]
        finalOpinion:
          type: ["string", "null"]
        initialSentiment:
          type: ["number", "null"]
        finalSentiment:
          type: ["number", "null"]
        beliefStance:
          type: ["number", "null"]
        beliefConfidence:
          type: ["number", "null"]
        keyConcern:
          type: ["string", "null"]
        wouldShare:
          type: ["boolean", "null"]
        platformPreference:
          type: ["string", "null"]
        voiceId:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        hasAudio:
          type: ["boolean", "null"]
        voiceScore:
          type: ["number", "null"]
        rank:
          type: ["integer", "null"]
        alignmentData:
          oneOf:
            - $ref: "#/components/schemas/AlignmentData"
            - type: "null"
      required:
        - id
        - analysisId

    ForecastPoint:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: string
        hourOffset:
          type: ["integer", "null"]
        sentiment:
          type: ["number", "null"]
        confidenceLow:
          type: ["number", "null"]
        confidenceHigh:
          type: ["number", "null"]
        isForecast:
          type: ["boolean", "null"]
      required:
        - id
        - analysisId

    ProblemSegment:
      type: object
      properties:
        quote:
          type: string
        triggeredBy:
          type: array
          items:
            type: string
        reason:
          type: string
      required:
        - quote
        - triggeredBy
        - reason

    MontageTimelineEntry:
      type: object
      properties:
        personaId:
          type: ["integer", "null"]
        personaName:
          type: string
        startSec:
          type: number
        endSec:
          type: number
        script:
          type: string
        words:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/AlignmentWord"
      required:
        - personaName
        - startSec
        - endSec
        - script

    SwarmSignedUrlResponse:
      type: object
      properties:
        signedUrl:
          type: string
      required:
        - signedUrl

    FullAnalysis:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        inputText:
          type: string
        status:
          type: string
        createdAt:
          type: string
        avgSentiment:
          type: ["number", "null"]
        dominantEmotion:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        viralPotential:
          type: ["number", "null"]
        consensusForming:
          type: ["boolean", "null"]
        swarmSummary:
          type: ["string", "null"]
        marketQuestion:
          type: ["string", "null"]
        marketProbability:
          type: ["number", "null"]
        factCheck:
          oneOf:
            - $ref: "#/components/schemas/FactCheckResult"
            - type: "null"
        exaResults:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/ExaResult"
        keyThemes:
          type: ["array", "null"]
          items:
            type: string
        narrativeFractures:
          type: ["array", "null"]
          items:
            type: string
        montageUrl:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        agentId:
          type: ["string", "null"]
        contentSuggestions:
          type: ["array", "null"]
          items:
            type: string
        problemSegments:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/ProblemSegment"
        montageTimeline:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/MontageTimelineEntry"
        personas:
          type: array
          items:
            $ref: "#/components/schemas/Persona"
        forecastPoints:
          type: array
          items:
            $ref: "#/components/schemas/ForecastPoint"
      required:
        - id
        - title
        - inputText
        - status
        - createdAt
        - personas
        - forecastPoints