# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwarmCast API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List recent analyses
 */
export const ListAnalysesResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
  avgSentiment: zod.number().nullish(),
  dominantEmotion: zod.string().nullish(),
  riskLevel: zod.string().nullish(),
});
export const ListAnalysesResponse = zod.array(ListAnalysesResponseItem);

/**
 * @summary Create a new analysis
 */
export const CreateAnalysisBody = zod.object({
  title: zod.string(),
  text: zod.string(),
});

/**
 * @summary Get full analysis result
 */
export const GetAnalysisParams = zod.object({
  id: zod.coerce.string(),
});

export const GetAnalysisResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  inputText: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
  avgSentiment: zod.number().nullish(),
  dominantEmotion: zod.string().nullish(),
  riskLevel: zod.string().nullish(),
  viralPotential: zod.number().nullish(),
  consensusForming: zod.boolean().nullish(),
  swarmSummary: zod.string().nullish(),
  marketQuestion: zod.string().nullish(),
  marketProbability: zod.number().nullish(),
  factCheck: zod
    .union([
      zod.object({
        objectiveAssessment: zod.string(),
        accuracyRating: zod.string(),
        factCheckItems: zod.array(
          zod.object({
            claim: zod.string(),
            status: zod.string(),
            note: zod.string(),
          }),
        ),
      }),
      zod.null(),
    ])
    .optional(),
  exaResults: zod
    .array(
      zod.object({
        title: zod.string(),
        url: zod.string(),
        snippet: zod.string(),
        publishedDate: zod.string().nullish(),
      }),
    )
    .nullish(),
  keyThemes: zod.array(zod.string()).nullish(),
  narrativeFractures: zod.array(zod.string()).nullish(),
  montageUrl: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  agentId: zod.string().nullish(),
  contentSuggestions: zod.array(zod.string()).nullish(),
  problemSegments: zod
    .array(
      zod.object({
        quote: zod.string(),
        triggeredBy: zod.array(zod.string()),
        reason: zod.string(),
      }),
    )
    .nullish(),
  montageTimeline: zod
    .array(
      zod.object({
        personaId: zod.number().nullish(),
        personaName: zod.string(),
        startSec: zod.number(),
        endSec: zod.number(),
        script: zod.string(),
        words: zod
          .array(
            zod.object({
              word: zod.string(),
              start: zod.number(),
              end: zod.number(),
            }),
          )
          .nullish(),
      }),
    )
    .nullish(),
  personas: zod.array(
    zod.object({
      id: zod.number(),
      analysisId: zod.string(),
      personaName: zod.string().nullish(),
      personaType: zod.string().nullish(),
      background: zod.string().nullish(),
      mbti: zod.string().nullish(),
      age: zod.number().nullish(),
      country: zod.string().nullish(),
      accent: zod.string().nullish(),
      gender: zod.string().nullish(),
      voiceStyle: zod.string().nullish(),
      activityLevel: zod.number().nullish(),
      influenceWeight: zod.number().nullish(),
      sentimentBias: zod.number().nullish(),
      initialReaction: zod.string().nullish(),
      finalOpinion: zod.string().nullish(),
      initialSentiment: zod.number().nullish(),
      finalSentiment: zod.number().nullish(),
      beliefStance: zod.number().nullish(),
      beliefConfidence: zod.number().nullish(),
      keyConcern: zod.string().nullish(),
      wouldShare: zod.boolean().nullish(),
      platformPreference: zod.string().nullish(),
      voiceId: zod.string().nullish(),
      audioUrl: zod.string().nullish(),
      hasAudio: zod.boolean().nullish(),
      voiceScore: zod.number().nullish(),
      rank: zod.number().nullish(),
      alignmentData: zod
        .union([
          zod.object({
            words: zod.array(
              zod.object({
                word: zod.string(),
                start: zod.number(),
                end: zod.number(),
              }),
            ),
          }),
          zod.null(),
        ])
        .optional(),
    }),
  ),
  forecastPoints: zod.array(
    zod.object({
      id: zod.number(),
      analysisId: zod.string(),
      hourOffset: zod.number().nullish(),
      sentiment: zod.number().nullish(),
      confidenceLow: zod.number().nullish(),
      confidenceHigh: zod.number().nullish(),
      isForecast: zod.boolean().nullish(),
    }),
  ),
});

/**
 * @summary SSE stream for analysis pipeline events
 */
export const StreamAnalysisParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get signed URL for SwarmCast ConvAI session
 */
export const GetSwarmSignedUrlParams = zod.object({
  id: zod.coerce.string(),
});

export const GetSwarmSignedUrlResponse = zod.object({
  signedUrl: zod.string(),
});