# lib/api-zod/src/generated/types/alignmentData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwarmCast API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlignmentWord } from "./alignmentWord";

export interface AlignmentData {
  words: AlignmentWord[];
}