# lib/api-zod/src/generated/types/fullAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwarmCast API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExaResult } from "./exaResult";
import type { FactCheckResult } from "./factCheckResult";
import type { ForecastPoint } from "./forecastPoint";
import type { MontageTimelineEntry } from "./montageTimelineEntry";
import type { Persona } from "./persona";
import type { ProblemSegment } from "./problemSegment";

export interface FullAnalysis {
  id: string;
  title: string;
  inputText: string;
  status: string;
  createdAt: string;
  /** @nullable */
  avgSentiment?: number | null;
  /** @nullable */
  dominantEmotion?: string | null;
  /** @nullable */
  riskLevel?: string | null;
  /** @nullable */
  viralPotential?: number | null;
  /** @nullable */
  consensusForming?: boolean | null;
  /** @nullable */
  swarmSummary?: string | null;
  /** @nullable */
  marketQuestion?: string | null;
  /** @nullable */
  marketProbability?: number | null;
  factCheck?: FactCheckResult | null;
  /** @nullable */
  exaResults?: ExaResult[] | null;
  /** @nullable */
  keyThemes?: string[] | null;
  /** @nullable */
  narrativeFractures?: string[] | null;
  /** @nullable */
  montageUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  contentSuggestions?: string[] | null;
  /** @nullable */
  problemSegments?: ProblemSegment[] | null;
  /** @nullable */
  montageTimeline?: MontageTimelineEntry[] | null;
  personas: Persona[];
  forecastPoints: ForecastPoint[];
}