# lib/api-zod/src/generated/types/persona.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwarmCast API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlignmentData } from "./alignmentData";

export interface Persona {
  id: number;
  analysisId: string;
  /** @nullable */
  personaName?: string | null;
  /** @nullable */
  personaType?: string | null;
  /** @nullable */
  background?: string | null;
  /** @nullable */
  mbti?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  accent?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  voiceStyle?: string | null;
  /** @nullable */
  activityLevel?: number | null;
  /** @nullable */
  influenceWeight?: number | null;
  /** @nullable */
  sentimentBias?: number | null;
  /** @nullable */
  initialReaction?: string | null;
  /** @nullable */
  finalOpinion?: string | null;
  /** @nullable */
  initialSentiment?: number | null;
  /** @nullable */
  finalSentiment?: number | null;
  /** @nullable */
  beliefStance?: number | null;
  /** @nullable */
  beliefConfidence?: number | null;
  /** @nullable */
  keyConcern?: string | null;
  /** @nullable */
  wouldShare?: boolean | null;
  /** @nullable */
  platformPreference?: string | null;
  /** @nullable */
  voiceId?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  hasAudio?: boolean | null;
  /** @nullable */
  voiceScore?: number | null;
  /** @nullable */
  rank?: number | null;
  alignmentData?: AlignmentData | null;
}