# lib/api-zod/src/generated/types/problemSegment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwarmCast API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProblemSegment {
  quote: string;
  triggeredBy: string[];
  reason: string;
}